chore: split binary and schema modules
