+++ docs/architecture.md
@@ -13,13 +13,13 @@ flowchart TD
 
     subgraph BINARY ["Binary decoding"]
         direction LR
-        B["stream.py\nouter frames"] --> C["reader.py\nbits & varints"]
+        B["binary/stream.py\nouter frames"] --> C["binary/reader.py\nbits & varints"]
     end
 
     subgraph SCHEMA ["Schema & state"]
         direction LR
-        D["sendtable.py\nserializer tree"] --> E["field_decoder.py\ntype dispatch"]
-        D --> F["field_path.py\nHuffman paths"]
+        D["schema/sendtable.py\nserializer tree"] --> E["schema/field_decoder.py\ntype dispatch"]
+        D --> F["schema/field_path.py\nHuffman paths"]
         G["string_table.py\nkey-value tables"]
         H["entities.py\ndelta updates"]
     end
@@ -76,11 +76,17 @@ flowchart TD
   <div class="arch-layer arch-layer--parse">
     <span class="arch-layer-label">Binary decoding</span>
     <div class="arch-layer-modules">
-      <span class="arch-badge">stream.py</span>
-      <span class="arch-badge">reader.py</span>
-      <span class="arch-badge">sendtable.py</span>
-      <span class="arch-badge">field_decoder.py</span>
-      <span class="arch-badge">field_path.py</span>
+      <span class="arch-badge">binary/stream.py</span>
+      <span class="arch-badge">binary/reader.py</span>
+    </div>
+  </div>
+
+  <div class="arch-layer arch-layer--parse">
+    <span class="arch-layer-label">Schema decoding</span>
+    <div class="arch-layer-modules">
+      <span class="arch-badge">schema/sendtable.py</span>
+      <span class="arch-badge">schema/field_decoder.py</span>
+      <span class="arch-badge">schema/field_path.py</span>
       <span class="arch-badge">string_table.py</span>
       <span class="arch-badge">entities.py</span>
     </div>
