+++ docs/reference/field_decoder.md
@@ -12,7 +12,7 @@ See also: [How Proto Parsing Works](../cookbook/proto-parsing-pipeline.md)
 
 ## Generated API
 
-## `gem.field_decoder.find_decoder`
+## `gem.schema.field_decoder.find_decoder`
 
 ### `find_decoder`
 
@@ -22,9 +22,9 @@ def find_decoder(field: _FieldLike) -> FieldDecoder
 
 Return the appropriate decoder for the given field.
 
-Source: [src/gem/field_decoder.py:403](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/field_decoder.py#L403)
+Source: [src/gem/schema/field_decoder.py:403](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L403)
 
-## `gem.field_decoder.find_decoder_by_base_type`
+## `gem.schema.field_decoder.find_decoder_by_base_type`
 
 ### `find_decoder_by_base_type`
 
@@ -34,9 +34,9 @@ def find_decoder_by_base_type(base_type: str) -> FieldDecoder
 
 Return a decoder for a base type string without field context.
 
-Source: [src/gem/field_decoder.py:435](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/field_decoder.py#L435)
+Source: [src/gem/schema/field_decoder.py:435](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L435)
 
-## `gem.field_decoder.QuantizedFloatDecoder`
+## `gem.schema.field_decoder.QuantizedFloatDecoder`
 
 ### `QuantizedFloatDecoder`
 
@@ -46,7 +46,7 @@ class QuantizedFloatDecoder
 
 Decoder for Source 2 quantized floats (CNetworkedQuantizedFloat).
 
-Source: [src/gem/field_decoder.py:66](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/field_decoder.py#L66)
+Source: [src/gem/schema/field_decoder.py:66](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L66)
 
 #### Methods
 
@@ -56,4 +56,4 @@ Signature: `def QuantizedFloatDecoder.decode(self, r: BitReader) -> float`
 
 Read and decode one quantized float from r.
 
-Source: [src/gem/field_decoder.py:180](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/field_decoder.py#L180)
+Source: [src/gem/schema/field_decoder.py:180](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L180)
