+++ docs/reference/sendtable.md
@@ -15,7 +15,7 @@ See also: [Proto Parsing Pipeline](../cookbook/proto-parsing-pipeline.md)
 
 ## Generated API
 
-## `gem.sendtable.parse_send_tables`
+## `gem.schema.sendtable.parse_send_tables`
 
 ### `parse_send_tables`
 
@@ -25,9 +25,9 @@ def parse_send_tables(data: bytes, game_build: int = 0) -> dict[str, Serializer]
 
 Parse a CDemoSendTables payload into a serializer dictionary.
 
-Source: [src/gem/sendtable.py:332](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L332)
+Source: [src/gem/schema/sendtable.py:332](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L332)
 
-## `gem.sendtable.Serializer`
+## `gem.schema.sendtable.Serializer`
 
 ### `Serializer`
 
@@ -37,7 +37,7 @@ class Serializer
 
 A named, versioned entity class schema with an ordered list of fields.
 
-Source: [src/gem/sendtable.py:223](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L223)
+Source: [src/gem/schema/sendtable.py:223](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L223)
 
 #### Dataclass fields
 
@@ -47,7 +47,7 @@ Source: [src/gem/sendtable.py:223](https://github.com/whanyu1212/gem-dota/blob/m
 | `version` | `int` | `-` |
 | `fields` | `list[Field]` | `field(...)` |
 
-## `gem.sendtable.Field`
+## `gem.schema.sendtable.Field`
 
 ### `Field`
 
@@ -57,7 +57,7 @@ class Field
 
 One property of a serializer, with its type and decoder.
 
-Source: [src/gem/sendtable.py:138](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L138)
+Source: [src/gem/schema/sendtable.py:138](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L138)
 
 #### Dataclass fields
 
@@ -89,7 +89,7 @@ Signature: `def Field.set_model(self, model: int) -> None`
 
 Assign the field model and wire up the appropriate decoders.
 
-Source: [src/gem/sendtable.py:179](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L179)
+Source: [src/gem/schema/sendtable.py:179](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L179)
 
 ##### `model_name`
 
@@ -97,9 +97,9 @@ Signature: `def Field.model_name(self) -> str`
 
 Return a human-readable model name for debugging.
 
-Source: [src/gem/sendtable.py:201](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L201)
+Source: [src/gem/schema/sendtable.py:201](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L201)
 
-## `gem.sendtable.FieldType`
+## `gem.schema.sendtable.FieldType`
 
 ### `FieldType`
 
@@ -109,7 +109,7 @@ class FieldType
 
 Parsed representation of a C++ field type string.
 
-Source: [src/gem/sendtable.py:76](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/sendtable.py#L76)
+Source: [src/gem/schema/sendtable.py:76](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/sendtable.py#L76)
 
 #### Dataclass fields
