+++ scripts/generate_vitepress_api_reference.py
@@ -35,14 +35,14 @@ TARGETS_BY_PAGE: dict[str, list[str]] = {
     "dataframes.md": ["gem.dataframes"],
     "entities.md": ["gem.entities.EntityOp", "gem.entities.Entity", "gem.entities.EntityManager"],
     "field_decoder.md": [
-        "gem.field_decoder.find_decoder",
-        "gem.field_decoder.find_decoder_by_base_type",
-        "gem.field_decoder.QuantizedFloatDecoder",
+        "gem.schema.field_decoder.find_decoder",
+        "gem.schema.field_decoder.find_decoder_by_base_type",
+        "gem.schema.field_decoder.QuantizedFloatDecoder",
     ],
     "field_path.md": [
-        "gem.field_path.read_field_paths",
-        "gem.field_path.FieldPath",
-        "gem.field_path.FieldPathOp",
+        "gem.schema.field_path.read_field_paths",
+        "gem.schema.field_path.FieldPath",
+        "gem.schema.field_path.FieldPathOp",
     ],
     "game_events.md": [
         "gem.game_events.GameEventManager",
@@ -52,14 +52,14 @@ TARGETS_BY_PAGE: dict[str, list[str]] = {
     "match_builder.md": ["gem.match_builder"],
     "models.md": ["gem.models"],
     "parser.md": ["gem.parser.ReplayParser"],
-    "reader.md": ["gem.reader.BitReader"],
+    "reader.md": ["gem.binary.reader.BitReader"],
     "sendtable.md": [
-        "gem.sendtable.parse_send_tables",
-        "gem.sendtable.Serializer",
-        "gem.sendtable.Field",
-        "gem.sendtable.FieldType",
+        "gem.schema.sendtable.parse_send_tables",
+        "gem.schema.sendtable.Serializer",
+        "gem.schema.sendtable.Field",
+        "gem.schema.sendtable.FieldType",
     ],
-    "stream.md": ["gem.stream.DemoStream", "gem.stream.OuterMessage"],
+    "stream.md": ["gem.binary.stream.DemoStream", "gem.binary.stream.OuterMessage"],
     "string_table.md": [
         "gem.string_table.StringTables",
         "gem.string_table.StringTable",
