+++ src/gem/binary/__init__.py
@@ -1,0 +1,6 @@
+"""Binary replay reading primitives."""
+
+from gem.binary.reader import BitReader, BufferError
+from gem.binary.stream import DemoStream, OuterMessage
+
+__all__ = ["BitReader", "BufferError", "DemoStream", "OuterMessage"]
