+++ src/gem/schema/__init__.py
@@ -1,0 +1,21 @@
+"""Replay schema and entity-field decoding primitives."""
+
+from gem.schema.field_decoder import FieldDecoder, QuantizedFloatDecoder
+from gem.schema.field_path import FieldPath, FieldPathOp, read_field_paths
+from gem.schema.field_reader import read_fields
+from gem.schema.field_state import FieldState
+from gem.schema.sendtable import Field, FieldType, Serializer, parse_send_tables
+
+__all__ = [
+    "Field",
+    "FieldDecoder",
+    "FieldPath",
+    "FieldPathOp",
+    "FieldState",
+    "FieldType",
+    "QuantizedFloatDecoder",
+    "Serializer",
+    "parse_send_tables",
+    "read_field_paths",
+    "read_fields",
+]
