+++ tests/test_fuzz.py
@@ -57,21 +57,21 @@ class TestDemoStreamFuzz:
     """DemoStream raises ValueError on bad magic; handles truncation cleanly."""
 
     def test_empty_bytes(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         # Empty bytes → wrong magic (0 bytes) → ValueError
         with pytest.raises(ValueError):
             DemoStream(b"")
 
     def test_wrong_magic(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         bad = b"NOTDOTA2" + b"\x00" * 8
         with pytest.raises(ValueError):
             DemoStream(bad)
 
     def test_header_only_no_messages(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         # Valid header, no messages — should yield nothing and not hang.
         with DemoStream(_make_dem_header()) as stream:
@@ -79,7 +79,7 @@ class TestDemoStreamFuzz:
         assert messages == []
 
     def test_truncated_after_header(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         # Header + incomplete varuint — should raise or yield nothing (no hang).
         data = _make_dem_header() + b"\x80"  # incomplete varuint
@@ -91,7 +91,7 @@ class TestDemoStreamFuzz:
                 pass  # any exception is fine — no hang
 
     def test_truncated_mid_payload(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         # Header + valid outer message claiming 100 bytes, only 10 provided.
         cmd = _pack_varuint32(7)  # DEM_Packet
@@ -108,7 +108,7 @@ class TestDemoStreamFuzz:
                 pass  # truncated payload raises — acceptable
 
     def test_garbage_after_magic(self) -> None:
-        from gem.stream import DemoStream
+        from gem.binary.stream import DemoStream
 
         # Valid magic, then garbage content — iteration may raise or return partial.
         data = _make_dem_header() + b"\xff" * 256
@@ -215,7 +215,7 @@ class TestBitReaderFuzz:
     """BitReader raises on reads past end-of-buffer."""
 
     def test_read_bits_past_end(self) -> None:
-        from gem.reader import BitReader
+        from gem.binary.reader import BitReader
 
         r = BitReader(b"\x00")
         r.read_bits(8)  # consume the single byte
@@ -223,14 +223,14 @@ class TestBitReaderFuzz:
             r.read_bits(1)  # past end
 
     def test_empty_buffer_raises(self) -> None:
-        from gem.reader import BitReader
+        from gem.binary.reader import BitReader
 
         r = BitReader(b"")
         with pytest.raises((EOFError, IndexError, ValueError, Exception)):
             r.read_bits(1)
 
     def test_read_byte_past_end(self) -> None:
-        from gem.reader import BitReader
+        from gem.binary.reader import BitReader
 
         r = BitReader(b"\xab")
         r.read_bits(8)  # consume all
