+++ tests/test_sendtable.py
@@ -1,5 +1,5 @@
 """
-Tests for gem.sendtable — send table parsing, FieldType, Field, Serializer.
+Tests for gem.schema.sendtable — send table parsing, FieldType, Field, Serializer.
 
 Reference: manta/sendtable.go, manta/field.go, manta/field_type.go
 """
@@ -14,7 +14,7 @@ import pytest
 class TestFieldTypeParsing:
     @pytest.fixture
     def parse(self):
-        from gem.sendtable import _parse_field_type
+        from gem.schema.sendtable import _parse_field_type
 
         return _parse_field_type
 
@@ -67,13 +67,13 @@ class TestFieldTypeParsing:
 class TestFieldModel:
     @pytest.fixture
     def field_cls(self):
-        from gem.sendtable import Field, FieldType
+        from gem.schema.sendtable import Field, FieldType
 
         return Field, FieldType
 
     def test_simple_model(self, field_cls):
         Field, FieldType = field_cls
-        from gem.sendtable import FIELD_MODEL_SIMPLE
+        from gem.schema.sendtable import FIELD_MODEL_SIMPLE
 
         f = Field(
             var_name="m_iHealth",
@@ -94,7 +94,7 @@ class TestFieldModel:
 
     def test_fixed_array_model(self, field_cls):
         Field, FieldType = field_cls
-        from gem.sendtable import FIELD_MODEL_FIXED_ARRAY
+        from gem.schema.sendtable import FIELD_MODEL_FIXED_ARRAY
 
         f = Field(
             var_name="m_hAbilities",
@@ -115,7 +115,7 @@ class TestFieldModel:
 
     def test_variable_array_model(self, field_cls):
         Field, FieldType = field_cls
-        from gem.sendtable import FIELD_MODEL_VARIABLE_ARRAY
+        from gem.schema.sendtable import FIELD_MODEL_VARIABLE_ARRAY
 
         f = Field(
             var_name="m_vecItems",
@@ -138,7 +138,7 @@ class TestFieldModel:
 
     def test_fixed_table_model(self, field_cls):
         Field, FieldType = field_cls
-        from gem.sendtable import FIELD_MODEL_FIXED_TABLE
+        from gem.schema.sendtable import FIELD_MODEL_FIXED_TABLE
 
         f = Field(
             var_name="m_pEntity",
@@ -166,7 +166,7 @@ class TestFieldModel:
 class TestFieldPatches:
     @pytest.fixture
     def patch_list(self):
-        from gem.sendtable import _FIELD_PATCHES
+        from gem.schema.sendtable import _FIELD_PATCHES
 
         return _FIELD_PATCHES
 
@@ -186,7 +186,7 @@ class TestFieldPatches:
             assert not p.should_apply(p.max_build + 1)
 
     def test_simtime_patch_sets_encoder(self, patch_list):
-        from gem.sendtable import Field, FieldType
+        from gem.schema.sendtable import Field, FieldType
 
         always_on = next(p for p in patch_list if p.min_build == 0 and p.max_build == 0)
         f = Field(
@@ -217,11 +217,11 @@ class TestParseSendTables:
         """Extract CDemoSendTables payload from the truncated fixture."""
         from pathlib import Path
 
+        from gem.binary.stream import DemoStream
         from gem.proto import (
             network_connection_pb2,  # noqa: F401
             networkbasetypes_pb2,  # noqa: F401
         )
-        from gem.stream import DemoStream
 
         fixture = Path(__file__).parent / "fixtures"
         dem_files = list(fixture.glob("*_truncated.dem"))
@@ -238,14 +238,14 @@ class TestParseSendTables:
         pytest.skip("CDemoSendTables not found in truncated fixture")
 
     def test_returns_nonempty_dict(self, sendtable_data):
-        from gem.sendtable import parse_send_tables
+        from gem.schema.sendtable import parse_send_tables
 
         serializers = parse_send_tables(sendtable_data)
         assert isinstance(serializers, dict)
         assert len(serializers) > 0
 
     def test_known_entity_classes_present(self, sendtable_data):
-        from gem.sendtable import parse_send_tables
+        from gem.schema.sendtable import parse_send_tables
 
         serializers = parse_send_tables(sendtable_data)
         # These classes appear in every Dota 2 replay
@@ -253,7 +253,7 @@ class TestParseSendTables:
             assert name in serializers, f"Missing expected serializer: {name}"
 
     def test_fields_have_decoders(self, sendtable_data):
-        from gem.sendtable import FIELD_MODEL_SIMPLE, parse_send_tables
+        from gem.schema.sendtable import FIELD_MODEL_SIMPLE, parse_send_tables
 
         serializers = parse_send_tables(sendtable_data)
         errors = []
@@ -264,7 +264,7 @@ class TestParseSendTables:
         assert not errors, f"Fields with no decoder: {errors[:5]}"
 
     def test_field_types_parsed(self, sendtable_data):
-        from gem.sendtable import parse_send_tables
+        from gem.schema.sendtable import parse_send_tables
 
         serializers = parse_send_tables(sendtable_data)
         for s in list(serializers.values())[:10]:
@@ -272,7 +272,7 @@ class TestParseSendTables:
                 assert f.field_type.base_type, f"Empty base_type for {s.name}.{f.var_name}"
 
     def test_serializer_repr(self, sendtable_data):
-        from gem.sendtable import parse_send_tables
+        from gem.schema.sendtable import parse_send_tables
 
         serializers = parse_send_tables(sendtable_data)
         s = next(iter(serializers.values()))
